feat(app-degree-pages): uds-903 externalise the main domain "webapp4.asu.edu"
I put the domain "webapp4.asu.edu" in a shared constant which can be used by the Degree component.
There are still a few place where that domain is present, but those are
only JSON file used for testing which do not really call that domain URL

I also did general clean up.