style: inactive colors for vscode
Update to help with context-switching. Background: UDS VSCode titleBar styles disappear when the window is inactive, unless defined like here. I've chosen to set the foreground text to a UDS-compatible gray when inactive so we retain a cue to inactivity while keeping the background of the titleBar a consistent maroon so the window is easily recognizable as a UDS window.